Поправка авторизации для синхронного обращения
